+++ src/lib.rs
@@ -18,6 +18,7 @@ pub mod failure_message;
 pub mod file_arg;
 pub mod filetype_report;
 pub mod free_space;
+pub mod gui;
 pub mod ini;
 pub mod log_eval;
 pub mod method_select;